move contest file to folder
